+++ _protected/views/trailers/_trailer.php
@@ -9,7 +9,8 @@ use yii\helpers\Html;
 	@var - $delivery_run_num - The delivery run this is for
 	@var - $target_delivery_load - the delivery load that his trailer slot sits in
 	@var - $trailer_slot_num - the trailer slot that this is being rendered for.
-	
+	@var = $delivery_load_id - the Delivery Load this trailer is to be rendered for
+	@var - $requestedDate - the date the delivery load is set for - used to get of list of bins already used
 	
 */
 
@@ -20,6 +21,17 @@ use yii\helpers\Html;
 * 
 */
 if(isset($trailer)) {  
+
+
+//get a list of the bins used//
+//$usedTrailerOtherLoads = $trailer->getUsedBinsOtherLoads($delivery_run_num, $requestedDate, $delivery_load_id);
+
+$deliveryLoadBins = Trailers::getDeliveryLoadBins($delivery_load_id);
+
+
+
+$otherDeliveryLoadsBins = Trailers::getUsedBinsOtherLoads($delivery_run_num, $requestedDate, $delivery_load_id);
+
 
 ?>
 <div class='trailer_display_<?= $trailer->id ?> trailer_details' style='width: 350px; margin-right: 30px; float:left;'
@@ -57,16 +69,48 @@ if(isset($trailer)) {
 			$count = 1;
 			foreach($trailer->trailerBins as  $trailerBin)
 				{
+					
+				//in case too many bins have been specified for the trailer
 				if($count > $trailer->NumBins)
 					{
 					exit;
 					}
 				
+				
+				//Trailer has been used in this delivery
+				if(array_key_exists($trailerBin->id, $deliveryLoadBins))
+					{
+					if($deliveryLoadBins[$trailerBin->id] < $trailerBin->MaxCapacity)
+						{
+						$class = 'sap_trailer_partial';
+						}
+					else{
+						$class = 'sap_trailer_full';
+						}	
+					echo "<div class='".$class."' style='width: ".$binDivWidth."%; border: 1px solid; height: 100%; float: left;  text-align:center;'>";
+					echo "Bin: ".$trailerBin->BinNo."<br>(".$trailerBin->MaxCapacity." T)<br>";					
+					echo "<input class='trailer_bin_checkbox trailer_cb_id_".$trailer->id."' trailer_id='".$trailer->id."' trailerbin_id='".$trailerBin->id."' capacity='".$trailerBin->MaxCapacity."' name='deliveryLoad[".$target_delivery_load."][truck_load][".$trailer->id."][".$trailerBin->id."][]' value='".$deliveryLoadBins[$trailerBin->id]."' checked type='checkbox' />";		
+					echo "</div>";	
+					}
+					
+				//Trailer bin is being used by another delivery load
+				elseif(array_key_exists($trailerBin->id, $otherDeliveryLoadsBins))
+					{
+					echo "<div class='sap_trailer_used' style='background-color: grey; width: ".$binDivWidth."%; border: 1px solid; height: 100%; float: left;  text-align:center;'>";
+					echo "<input type='hidden' class='trailer_cb_id_".$trailer->id."' value='1'>";
+					echo "Bin: ".$trailerBin->BinNo."<br>(".$trailerBin->MaxCapacity." T)<br>";
+					
+					echo "</div>";	
+					}
 				//Trailer bin hasn't been used in this delivery or any other delviery.
-				echo "<div class='sap_trailer_empty' style='width: ".$binDivWidth."%; border: 1px solid; height: 100%; float: left;  text-align:center;'>";
-				echo "Bin: ".$trailerBin->BinNo."<br>(".$trailerBin->MaxCapacity." T)<br>";
-				echo "<input class='trailer_bin_checkbox trailer_cb_id_".$trailer->id."' trailer_id='".$trailer->id."' trailerbin_id='".$trailerBin->id."' capacity='".$trailerBin->MaxCapacity."' name='deliveryLoad[".$target_delivery_load."][truck_load][".$trailer->id."][".$trailerBin->id."][]' value='0' type='checkbox' />";	
-				echo "</div>";
+				else{
+					echo "<div class='sap_trailer_empty' style='width: ".$binDivWidth."%; border: 1px solid; height: 100%; float: left;  text-align:center;'>";
+					echo "Bin: ".$trailerBin->BinNo."<br>(".$trailerBin->MaxCapacity." T)<br>";
+					echo "<input class='trailer_bin_checkbox trailer_cb_id_".$trailer->id."' trailer_id='".$trailer->id."' trailerbin_id='".$trailerBin->id."' capacity='".$trailerBin->MaxCapacity."' name='deliveryLoad[".$target_delivery_load."][truck_load][".$trailer->id."][".$trailerBin->id."][]' value='0' type='checkbox' />";	
+					echo "</div>";
+					}
+				
+
